ubah login redirect dari home ke admin

+++ routes/web.php
@@ -24,16 +24,12 @@ Route::post('ckeditor/image_upload', 'CKEditorController@upload')->name('upload'
 
 Route::group(['middleware' => 'auth'], function () {
 
-    Route::get('/home', 'HomeController@index')->name('home');
-    Route::resource('/category', 'CategoryController');
-    Route::resource('/tag', 'TagController');
+    Route::get('/admin', 'HomeController@index')->name('admin');
+    Route::resource('/admin/category', 'CategoryController');
+    Route::resource('/admin/tag', 'TagController');
 
-    Route::get('/post/tampil_hapus', 'PostController@tampil_hapus')->name('post.tampil_hapus');
-    Route::get('/post/restore/{id}', 'PostController@restore')->name('post.restore');
-    Route::delete('/post/kill/{id}', 'PostController@kill')->name('post.kill');
-    Route::resource('/post', 'PostController');
-
-    Route::middleware(['auth:sanctum', 'verified'])->get('/dashboard', function () {
-        return view('dashboard');
-    })->name('dashboard');
+    Route::get('/admin/post/tampil_hapus', 'PostController@tampil_hapus')->name('post.tampil_hapus');
+    Route::get('/admin/post/restore/{id}', 'PostController@restore')->name('post.restore');
+    Route::delete('/admin/post/kill/{id}', 'PostController@kill')->name('post.kill');
+    Route::resource('/admin/post', 'PostController');
 });
